# Small Encoding Performance Improvement